[font3] gxui: use gxfont.Default, remove loading by name
This leaves the discovery and bundling of fonts up to the user's
application. A default is provided, built into their binary with the
default theme (meaning a single binary can still be distributed).

The idea underlying this change is GXUI strives to be system independent.
Leave dependent things like finding fonts up to applications that
understand their systems.

Change-Id: Ic2eb14b1846e0a6650791a8e3f5988505009071c