gxui/gxfont: add a fixed-width font
For some reason I thought the copyright on Droid Sans Mono was held by
another organization, but it appears to be owned by Google. So it makes
an easy inclusion.

The gxfont.a file grows to 100kb with the extra font. This isn't great,
but right now the size is dwarfed by gl/glfw which brings simple
binaries into the 8mb range.