Remove panics for rendering a destroyed viewport.
The panic was there to sanity-check that the application was not attempting
to update the canvas after tear-down. However, in this case the tear-down
was started because of the window being closed which is logic external to
the app-logic. The close message gets put on the driver event queue making
it near impossible for the application to know whether it is safe to call
SetCanvas or not while the panic test is there.

The logic for checking whether the viewport is alive and the render should
happen seems sound, so instead of panicing, make the render a no-op on
a destroyed viewport.

Fixes: #77