Add a new ContainsItem 'virtual' method to mixins.List.
This is used to decide whether an item is part of the adapter.
This is slightly complicated as Trees should report true for items
that are not part of the expanded set, which is completely hidden
from what the List sees. To solve, override ContainsItem in Tree to
inspect the TreeAdapter's full set of items, not just the expanded
set.

Fixes #80