Fix for Trees not updating their counts resulting in hidden items.
The adjustDescendants call was sometimes not bubbling up to the adapter.
This was due to the parent becoming incorrect for the root nodes’s
children when the root was updated.

Removed the anonymous field embedding of TreeToListNode in
TreeToListAdapter to try and avoid this kind of bug in the future.